+++ src/main/java/net/whydah/sso/session/baseclasses/BaseWhydahServiceClient.java
@@ -160,7 +160,7 @@ public class BaseWhydahServiceClient {
     }
 
     public static Integer calculateTokenRemainingLifetimeInSeconds(String userTokenXml) {
-        Integer tokenLifespanSec = UserTokenXpathHelper.getLifespan(userTokenXml);
+        Long tokenLifespanSec = UserTokenXpathHelper.getLifespan(userTokenXml);
         Long tokenTimestampMsSinceEpoch = UserTokenXpathHelper.getTimestamp(userTokenXml);
 
         if (tokenLifespanSec == null || tokenTimestampMsSinceEpoch == null) {
